add style to match the solution
I did everything looking the same but with some differences to solution.

I did add display:flex to the item which was not neccessary.
I was supposed to add this only to container and set max-width for item only.
I have done it because wanted to center these items forgetting I can use text-align in body section.